Force use of const colorspace and profile pointers
This removes the usage of grabColorspace/releaseColorspace from Krita.

The option to get a non-const colorspace existed to get a non-const
profile, on which the current exposure could be set. That changed with
the move to lcms-based hdr colorspaces, so it's no longer necessary.

+++ krita/plugins/paintops/libpaintop/kis_color_source.cpp
@@ -114,7 +114,7 @@ void KisPlainColorSource::selectColor(double mix)
     colors[0] = m_cachedBackGroundColor->data();
     colors[1] = m_foreGroundColor.data();
     int weight = (int)(mix * 255);
-    const qint16 weights[2] = { 255 - weight, weight };
+    const qint16 weights[2] = { (qint16)255 - weight, (qint16)weight };
 
     m_color->colorSpace()->mixColorsOp()->mixColors(colors, weights, 2, m_color->data());
